libnm-util: clarify/fix the definition of NM_SETTING_COMPARE_FLAG_FUZZY
Clarify that NM_SETTING_COMPARE_FLAG_FUZZY means "match only
attributes that affect how or whether the setting is applied to a
device", and update the use of NM_SETTING_PARAM_FUZZY_IGNORE to match.

This means that NMSettingWireless:seen-bssids is now the only
FUZZY_IGNORE property outside of NMSettingConnection, and
NMSettingConnection:autoconnect is now not fuzzy (since changing it
could cause the connection to be activated). Also,
NMSettingConnection:master, :slave-type, and :secondaries are now
non-fuzzy, which they probably were supposed to have been before.