core: ensure that ifname of NMDevice is valid
The device object needs a valid ifname, assert against it.
This name is also used to construct path names, so this checks
that the ifname/ip_ifname is not "." nor "..", and that it does
not contain '/'.

This check is not expected to ever fail, and if it does, it hints to a
bug somewhere else. If this assert fails, the device is left in an
unusable state and NM might crash later on.

The check is more relaxed then nm_utils_iface_valid_name, because
we only want to make sure, that constructing path names is not
dangerous. We don't have the strong requirements on the interface name
from the kernel, because the iface name might be e.g. bluez bdaddr.
Especially, we don't limit the interface name to 15 chars.

Signed-off-by: Thomas Haller <thaller@redhat.com>