libnm-glib: fix crash when NMVPNPlugin unrefs self in constructor (rh #1050934)
When NMVPNPlugin encountered an error at construct time, it would unref itself
and return NULL. glib 2.39 will g_error() out in this case, because you're not
supposed to do that, and because it used to cause mysterious crashes later on
if you did.
Fix this by using GInitable to initialize NMVPNPlugin rather than trying to do
it from constructor().

See also
  https://mail.gnome.org/archives/commits-list/2013-October/msg07491.html

https://bugzilla.redhat.com/show_bug.cgi?id=1050934