config: add new NMConfigData class
A NMConfig instance is immutable and it's properties do not change,
with one exception: the NMConfigData property.

The NMConfigData instance itself is also immutable and does not change,
with one exception: the NMConfig parent reference will be set to NULL
when the config gets disposed. NMConfigData only holds a weak reference
to the NMConfig.

When reloading a configuration, NMConfig will only swap the NMConfigData
instance.