config: add new NMConfigData class
In the future, the NMConfig class should be immutable and its
properties should not change, with one exception: the NMConfigData property.

The NMConfigData instance itself is also immutable and does not change,
with one exception: the NMConfig parent reference will be set to %NULL
when the parent config gets disposed. NMConfigData only holds a weak reference
to the NMConfig.

Later, when changing/reloading a configuration, NMConfig will only swap the
NMConfigData instance.