connectivity: fix race checking connectivity when setting online
When setting the state to 'online' we always have to
run a new check. Otherwise, we might not schedule a new check
because of @running. However, the currently running check might
fail (because it was started before, when we had no connectivity
yet).

The optimal solution would remember that we have to run a check
and place a limit of how many checks we perform.

Signed-off-by: Thomas Haller <thaller@redhat.com>