core: drop "default-unmanaged" devices
"default-unmanaged" was added for NMDeviceGeneric, so that they could
be unmanaged by default, but still activatable under the correct
arcane circumstances. However, NMDeviceGeneric will only allow itself
connections with type "generic" to be activated on it anyway, so there
is no chance of a device being accidentally activated on it, so all of
the unmanaged-default handling just complicates the code for no
reason.