tui: miscellaneous NmtNewt API improvements
Treat setting NmtNewtEntry:text to NULL as though it had been set to
"", since it may get bound to a property which is sometimes NULL.

Add NmtNewtListbox functions to get the key associated with a row, and
to append a row with a GDestroyNotify for its key.

Allow creating an empty NmtNewtPopup and then filling it in afterward.