dnsmasq: kill running process from pidfile synchronously
This should avoid a race when stopping dnsmasq.

This is still a hack, because we don't want to ever wait
for a process to terminate. Also, this will keep NM hanging
if the process does not die on SIGKILL (which however should
only happen when something is seriously wrong already).

A better solution would be to have nm-dnsmasq-manager managing
the global state. Not only the dnsmasq instance of on NMDevice.
Instead it should keep track of all the child processes it starts
and it wants to start. This way, starting a new process should be
delayed until any (child or non-child) instances are gone.
That however would be a bigger change.

https://bugzilla.gnome.org/show_bug.cgi?id=728342

Signed-off-by: Thomas Haller <thaller@redhat.com>