dns: kill plugin process synchronously
This should avoid a race when stopping dnsmasq.

This is still a hack, because we don't want to ever wait
for a process to terminate. Also, this will keep NM hanging
if the process does not die on SIGKILL (which however should
only happen when something is seriously wrong already).

https://bugzilla.gnome.org/show_bug.cgi?id=728342