core: don't assume connections for INTERNAL or PARENT unmanaged devices
INTERNAL is actually a nop right now because the only thing that
sets it is suspend/resume, which is covered by the preceding
manager_sleeping() call.  But we may use this more in the future,
so add it while we're here.

Devices that are unmanaged because their parent is unmanaged
probably shouldn't assume connections either, per 4e105c50.

(cherry picked from commit 71027ce5f1cec2f3356deb1edaf9d0d346a9e067)