build: embed git-commit-id as string inside the compiled libnm
In the 'configure.ac' script we already detect the git commit id
for the current source version. When creating a tarball, it is also
included inside the generated 'configure' script.

Add the commit id as a  static string to nm-utils.c. That way, having
a build of libnm.so or NetworkManager, you can quickly find the
corresponding git commit:

    strings src/NetworkManager | grep NM_GIT_SHA

Note that this only works after a new autogen.sh. Simply rebuilding
is not enough. Hence, to ensure that you have the right commit id
embedded, rebuild cleanly.